A bunch of compile cleanups. StdString.h is left [uggh]

+++ Musik.MFC/CSizingControlBar/sizecbar.cpp
@@ -254,8 +254,8 @@ CSize CSizingControlBar::CalcFixedLayout(BOOL bStretch, BOOL bHorz)
     GetRowSizingBars(arrSCBars);
     AFX_SIZEPARENTPARAMS layout;
     layout.hDWP = pDockBar->m_bLayoutQuery ?
-        NULL : ::BeginDeferWindowPos(arrSCBars.GetSize());
-    for (int i = 0; i < arrSCBars.GetSize(); i++)
+        NULL : ::BeginDeferWindowPos((int)arrSCBars.GetSize());
+    for (int i = 0; i < (int)arrSCBars.GetSize(); i++)
         if (arrSCBars[i]->m_nStateFlags & (delayHide|delayShow))
             arrSCBars[i]->RecalcDelayShow(&layout);
     if (layout.hDWP != NULL)
@@ -755,7 +755,7 @@ void CSizingControlBar::OnTrackUpdateSize(CPoint& point)
 
         // the others are shrinking
         int nFirst = bBefore ? nGrowingBar - 1 : nGrowingBar + 1;
-        int nLimit = bBefore ? -1 : arrSCBars.GetSize();
+        int nLimit = bBefore ? -1 : (int)arrSCBars.GetSize();
 
         for (int i = nFirst; nDelta != 0 && i != nLimit; i += (bBefore ? -1 : 1))
         {
@@ -864,7 +864,7 @@ void CSizingControlBar::GetRowInfo(int& nFirst, int& nLast, int& nThis)
     nThis = m_pDockBar->FindBar(this);
     ASSERT(nThis != -1);
 
-    int i, nBars = m_pDockBar->m_arrBars.GetSize();
+    int i, nBars = (int)m_pDockBar->m_arrBars.GetSize();
 
     // find the first and the last bar in row
     for (nFirst = -1, i = nThis - 1; i >= 0 && nFirst == -1; i--)
@@ -900,7 +900,7 @@ void CSizingControlBar::GetRowSizingBars(CSCBArray& arrSCBars, int& nThis)
         if (pBar->IsKindOf(RUNTIME_CLASS(CSizingControlBar)))
         {
             if (pBar == this)
-                nThis = arrSCBars.GetSize();
+                nThis = (int)arrSCBars.GetSize();
 
             arrSCBars.Add(pBar);
         }
@@ -980,7 +980,7 @@ BOOL CSizingControlBar::NegotiateSpace(int nLengthTotal, BOOL bHorz)
 
     CSCBArray arrSCBars;
     GetRowSizingBars(arrSCBars);
-    int nNumBars = arrSCBars.GetSize();
+    int nNumBars = (int)arrSCBars.GetSize();
     int nDelta = nLengthAvail - nLengthActual;
 
     // return faster when there is only one sizing bar per row (this one)
